+++ modules/core/src/test/java/org/apache/ignite/internal/processors/cache/distributed/near/GridCacheNearMultiGetSelfTest.java
@@ -83,7 +83,7 @@ public class GridCacheNearMultiGetSelfTest extends GridCommonAbstractTest {
 
         spi.setIpFinder(ipFinder);
 
-        c.setFailureDetectionTimeout(Long.MAX_VALUE);
+        c.setFailureDetectionTimeout(Integer.MAX_VALUE);
         c.setDiscoverySpi(spi);
 
         c.setCacheConfiguration(cc);
